# services/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "SSDB API",
    "version" : "1.0.0"
  },
  "paths" : {
...
    "/management/update-restaurant-infos" : {
      "post" : {
        "summary" : "updateRestaurantInfos",
        "operationId" : "updateRestaurantInfos",
        "parameters" : [ {
          "name" : "restaurant-id",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
    "/management/update-dish" : {
      "put" : {
        "summary" : "updateDish",
        "operationId" : "updateDish",
        "parameters" : [ {
          "name" : "restaurant-id",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
    "/management/add-dish" : {
      "post" : {
        "summary" : "addDish",
        "operationId" : "addDish",
        "parameters" : [ {
          "name" : "restaurant-id",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
    "/management/remove-dish" : {
      "delete" : {
        "summary" : "removeDish",
        "operationId" : "removeDish",
        "parameters" : [ {
          "name" : "restaurant-id",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dish-id",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
...
    "/group-orders" : {
      "post" : {
        "summary" : "createGroupOrder",
        "operationId" : "createGroupOrder",
        "parameters" : [ {
          "name" : "delivery-location-id",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "delivery-date-time",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
...
  },
...
}